pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "my-profile-pipeline"
version = "0.1.0"
description = "AI ingestion pipeline for my-profile"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "feedparser>=6.0",
    "httpx>=0.27",
    "readability-lxml>=0.8",
    "lxml>=5.0",
    "yt-dlp>=2024.1.1",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-httpx>=0.30",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pipeline*"]
